pyproject.toml
[project]
name = "long-text-agent"
version = "0.1.0"
description = "长文本生成策略基础Agent"
readme = "ReadMe.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv==1.0.1",
  "requests==2.32.3",
]
